// packages/data-ingest/source/types.ts
export interface Fact {
	sequence?: number;
	language_uid?: number;
	language?: string;
	lh_context_uid?: number;
	lh_context_name?: string;
	lh_object_uid: number;
	lh_cardinalities?: string;
	lh_object_name?: string;
	fact_uid: number;
	rel_type_uid: number;
	rel_type_name?: string;
	rh_object_uid: number;
	rh_cardinalities?: string;
	rh_object_name?: string;
	partial_definition?: string;
	full_definition?: string;
	uom_uid?: number;
	uom_name?: string;
	remarks?: string;
	approval_status?: string;
	successor_uid?: number;
	effective_from?: string;
	latest_update?: string;
	author?: string;
	reference?: string;
	collection_uid?: number;
	collection_name?: string;
	validity_context_uid?: number;
	validity_context_name?: string;
}